#374 refactor untilities
