+++ src/opera/pge/cslc_s1/cslc_s1_pge.py
@@ -21,8 +21,8 @@ from opera.pge.base.base_pge import PgeExecutor
 from opera.pge.base.base_pge import PostProcessorMixin
 from opera.pge.base.base_pge import PreProcessorMixin
 from opera.util.error_codes import ErrorCode
+from opera.util.h5_utils import get_cslc_s1_product_metadata
 from opera.util.input_validation import validate_slc_s1_inputs
-from opera.util.metadata_utils import get_cslc_s1_product_metadata
 from opera.util.render_jinja2 import render_jinja2
 from opera.util.time import get_time_for_filename
