+++ src/opera/test/data/test_cslc_s1_static_config.yaml
@@ -28,7 +28,7 @@ RunConfig:
                 ProgramPath: mkdir
                 ProgramOptions:
                     - '-p cslc_pge_test/output_dir/t064_135518_iw1/20220501/;'
-                    - 'python3 -c "from opera.util.metadata_utils import create_test_cslc_metadata_product; create_test_cslc_metadata_product(\"cslc_pge_test/output_dir/t064_135518_iw1/20220501/static_layers_t064_135518_iw1_20220501.h5\")";'
+                    - 'python3 -c "from opera.util.h5_utils import create_test_cslc_metadata_product; create_test_cslc_metadata_product(\"cslc_pge_test/output_dir/t064_135518_iw1/20220501/static_layers_t064_135518_iw1_20220501.h5\")";'
                     - '/bin/echo CSLC-S1 invoked with RunConfig'
                 ErrorCodeBase: 200000
                 SchemaPath: pge/cslc_s1/schema/cslc_s1_sas_schema.yaml
