+++ src/opera/test/data/test_rtc_s1_config.yaml
@@ -25,7 +25,7 @@ RunConfig:
                 ProgramPath: mkdir
                 ProgramOptions:
                     - '-p rtc_s1_test/output_dir/t069_147170_iw1/;'
-                    - 'python3 -c "from opera.util.metadata_utils import create_test_rtc_metadata_product; create_test_rtc_metadata_product(\"rtc_s1_test/output_dir/t069_147170_iw1/rtc_product_v1.0.h5\")";'
+                    - 'python3 -c "from opera.util.h5_utils import create_test_rtc_metadata_product; create_test_rtc_metadata_product(\"rtc_s1_test/output_dir/t069_147170_iw1/rtc_product_v1.0.h5\")";'
                     - 'dd if=/dev/urandom of=rtc_s1_test/output_dir/t069_147170_iw1/OPERA_L2_RTC-S1_T069-147170-IW1_20180504T104508Z_20230911T201937Z_S1B_30_v1.0_VH.tif bs=1M count=1;'
                     - 'dd if=/dev/urandom of=rtc_s1_test/output_dir/t069_147170_iw1/OPERA_L2_RTC-S1_T069-147170-IW1_20180504T104508Z_20230911T201937Z_S1B_30_v1.0_VV.tif bs=1M count=1;'
                     - 'dd if=/dev/urandom of=rtc_s1_test/output_dir/t069_147170_iw1/OPERA_L2_RTC-S1_T069-147170-IW1_20180504T104508Z_20230911T201937Z_S1B_30_v1.0_HH.tif bs=1M count=1;'
