+++ src/opera/test/pge/dswx_s1/test_dswx_s1_pge.py
@@ -23,8 +23,8 @@ import yaml
 from opera.pge import RunConfig
 from opera.pge.dswx_s1.dswx_s1_pge import DSWxS1Executor
 from opera.util import PgeLogger
-from opera.util.img_utils import MockGdal
-from opera.util.metadata_utils import get_sensor_from_spacecraft_name
+from opera.util.dataset_utils import get_sensor_from_spacecraft_name
+from opera.util.tiff_utils import MockGdal
 
 
 class DswxS1PgeTestCase(unittest.TestCase):
