+++ src/opera/test/pge/rtc_s1/test_rtc_s1_pge.py
@@ -22,15 +22,15 @@ from pkg_resources import resource_filename
 
 import yaml
 
-import opera.util.img_utils
+import opera.util.tiff_utils
 from opera.pge import RunConfig
 from opera.pge.rtc_s1.rtc_s1_pge import RtcS1Executor
 from opera.util import PgeLogger
-from opera.util.img_utils import mock_gdal_edit
-from opera.util.img_utils import mock_save_as_cog
-from opera.util.metadata_utils import create_test_rtc_metadata_product
-from opera.util.metadata_utils import get_rtc_s1_product_metadata
-from opera.util.metadata_utils import get_sensor_from_spacecraft_name
+from opera.util.dataset_utils import get_sensor_from_spacecraft_name
+from opera.util.h5_utils import create_test_rtc_metadata_product
+from opera.util.h5_utils import get_rtc_s1_product_metadata
+from opera.util.tiff_utils import mock_gdal_edit
+from opera.util.tiff_utils import mock_save_as_cog
 
 
 class RtcS1PgeTestCase(unittest.TestCase):
@@ -93,8 +93,8 @@ class RtcS1PgeTestCase(unittest.TestCase):
         os.chdir(self.test_dir)
         self.working_dir.cleanup()
 
-    @patch.object(opera.util.img_utils, "gdal_edit", mock_gdal_edit)
-    @patch.object(opera.util.img_utils, "save_as_cog", mock_save_as_cog)
+    @patch.object(opera.util.tiff_utils, "gdal_edit", mock_gdal_edit)
+    @patch.object(opera.util.tiff_utils, "save_as_cog", mock_save_as_cog)
     def test_rtc_s1_pge_execution(self):
         """
         Test execution of the RtcS1Executor class and its associated mixins
@@ -156,7 +156,7 @@ class RtcS1PgeTestCase(unittest.TestCase):
 
         self.assertIn(f"RTC-S1 invoked with RunConfig {expected_sas_config_file}", log_contents)
 
-    @patch.object(opera.util.img_utils, "save_as_cog", mock_save_as_cog)
+    @patch.object(opera.util.tiff_utils, "save_as_cog", mock_save_as_cog)
     def test_static_layer_data_access_url_injection(self):
         """Test injection of static data access URL into the output HDF5 product(s)"""
         expected_url = 'https://search.asf.alaska.edu/#/?dataset=OPERA-S1&productTypes=RTC-STATIC&operaBurstID=T069-147170-IW1&end=2018-05-04'   # noqa E501
@@ -176,7 +176,7 @@ class RtcS1PgeTestCase(unittest.TestCase):
             self.assertTrue(args[3].endswith('.tif'))
             return 0
 
-        with patch.object(opera.util.img_utils, "gdal_edit", _patched_gdal_edit):
+        with patch.object(opera.util.tiff_utils, "gdal_edit", _patched_gdal_edit):
             pge.run()
 
         # Grab the metadata generated from the PGE run
@@ -191,8 +191,8 @@ class RtcS1PgeTestCase(unittest.TestCase):
         self.assertIn('staticLayersDataAccess', rtc_metadata['identification'])
         self.assertEqual(rtc_metadata['identification']['staticLayersDataAccess'], expected_url)
 
-    @patch.object(opera.util.img_utils, "gdal_edit", mock_gdal_edit)
-    @patch.object(opera.util.img_utils, "save_as_cog", mock_save_as_cog)
+    @patch.object(opera.util.tiff_utils, "gdal_edit", mock_gdal_edit)
+    @patch.object(opera.util.tiff_utils, "save_as_cog", mock_save_as_cog)
     def test_filename_application(self):
         """Test the filename convention applied to RTC output products"""
         runconfig_path = join(self.data_dir, 'test_rtc_s1_config.yaml')
@@ -250,8 +250,8 @@ class RtcS1PgeTestCase(unittest.TestCase):
 
         self.assertIn(f"{core_filename}_BROWSE.png", output_files)
 
-    @patch.object(opera.util.img_utils, "gdal_edit", mock_gdal_edit)
-    @patch.object(opera.util.img_utils, "save_as_cog", mock_save_as_cog)
+    @patch.object(opera.util.tiff_utils, "gdal_edit", mock_gdal_edit)
+    @patch.object(opera.util.tiff_utils, "save_as_cog", mock_save_as_cog)
     def test_static_layer_filename_application(self):
         """Test the filename convention applied to RTC static layer output products"""
         runconfig_path = join(self.data_dir, 'test_rtc_s1_static_config.yaml')
@@ -601,8 +601,8 @@ class RtcS1PgeTestCase(unittest.TestCase):
             if os.path.exists(test_runconfig_path):
                 os.unlink(test_runconfig_path)
 
-    @patch.object(opera.util.img_utils, "gdal_edit", mock_gdal_edit)
-    @patch.object(opera.util.img_utils, "save_as_cog", mock_save_as_cog)
+    @patch.object(opera.util.tiff_utils, "gdal_edit", mock_gdal_edit)
+    @patch.object(opera.util.tiff_utils, "save_as_cog", mock_save_as_cog)
     def test_no_burst_id(self):
         """Test _iso_metadata_filename with bad burst_id"""
         runconfig_path = join(self.data_dir, 'test_rtc_s1_config.yaml')
@@ -612,8 +612,8 @@ class RtcS1PgeTestCase(unittest.TestCase):
         with self.assertRaises(RuntimeError):
             pge._iso_metadata_filename("bad_burst")
 
-    @patch.object(opera.util.img_utils, "gdal_edit", mock_gdal_edit)
-    @patch.object(opera.util.img_utils, "save_as_cog", mock_save_as_cog)
+    @patch.object(opera.util.tiff_utils, "gdal_edit", mock_gdal_edit)
+    @patch.object(opera.util.tiff_utils, "save_as_cog", mock_save_as_cog)
     def test_bad_iso_metadata_template(self):
         """Verify code when the QA application is enabled"""
         runconfig_path = join(self.data_dir, 'test_rtc_s1_config.yaml')
@@ -645,8 +645,8 @@ class RtcS1PgeTestCase(unittest.TestCase):
             if os.path.exists(test_runconfig_path):
                 os.unlink(test_runconfig_path)
 
-    @patch.object(opera.util.img_utils, "gdal_edit", mock_gdal_edit)
-    @patch.object(opera.util.img_utils, "save_as_cog", mock_save_as_cog)
+    @patch.object(opera.util.tiff_utils, "gdal_edit", mock_gdal_edit)
+    @patch.object(opera.util.tiff_utils, "save_as_cog", mock_save_as_cog)
     def test_qa_enabled(self):
         """Test the staging of the qa.log files."""
         # Verify code when the QA application is enabled
